loft-taxi week-1 corrections
