+++ src/index.js
@@ -3,11 +3,13 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
+import { theme } from 'loft-taxi-mui-theme';
+import { MuiThemeProvider } from '@material-ui/core/styles'
 
 ReactDOM.render(
-  <React.StrictMode>
+  <MuiThemeProvider theme={ theme }>
     <App />
-  </React.StrictMode>,
+  </MuiThemeProvider>,
   document.getElementById('root')
 );
