--- src/pgs/PageProfile.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export const PageProfile = () => {
-    return (
-        <div>PageProfile</div>
-    )
-}
